refactor: location module renamed to locations
