+++ src/lib/locations/locations.go
@@ -1,4 +1,4 @@
-package location
+package locations
 
 import (
 	"appengine"
