+++ src/lib/locations/select.go
@@ -1,4 +1,4 @@
-package location
+package locations
 
 import (
 	"appengine"
